pyproject.toml
[build-system]
requires = ["hatchling>=1.25.0"]
build-backend = "hatchling.build"

[project]
name = "ankiday"
version = "0.1.0"
description = "CLI to manage Anki decks, note types, and notes from YAML config (Anki - Deck as YAML)"
authors = [
  { name = "Boris Kukushkin" }
]
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
keywords = ["anki", "ankiconnect", "flashcards", "cli", "yaml", "deck-as-yaml"]
dependencies = [
  "typer[all]>=0.12.0",
  "pydantic>=2.8.0",
  "pyyaml>=6.0.1",
  "httpx>=0.27.0"
]

[project.scripts]
ankiday = "ankiday.cli:app"

[tool.hatch.build.targets.wheel]
packages = ["ankiday"]

[tool.ruff]
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP", "SIM"]

[tool.ruff.format]
quote-style = "double"
